# Improve vectorization in get_sad