This is part of a series of commits authored by @maj160 to improve performance of rav1e.

This commit results in an overall performance improvement of about 1% at speed 2 10-bit:

```
Benchmark 1: ~/Downloads/rav1e_master -s 2 --quantizer 64 ~/xiph-media-files/objective-1-fast-10bit/speed_bag_640x360_60f.y4m -y -o /dev/null
  Time (mean ± σ):     43.455 s ±  0.134 s    [User: 43.464 s, System: 0.179 s]
  Range (min … max):   43.305 s … 43.705 s    10 runs

Benchmark 2: ~/Downloads/rav1e_mod -s 2 --quantizer 64 ~/xiph-media-files/objective-1-fast-10bit/speed_bag_640x360_60f.y4m -y -o /dev/null
  Time (mean ± σ):     42.926 s ±  0.098 s    [User: 42.945 s, System: 0.166 s]
  Range (min … max):   42.783 s … 43.097 s    10 runs
```